Fix mypy CI failures: variable-shadow narrowing, numba dispatch typing, and wire up real fastmath
Renamed reused loop/branch variables in camera.py and flock.py so mypy
stops unifying incompatible array shapes/dtypes across separate code
paths. Added explicit float32 casts in boid_init.py/influencer.py where
an intermediate op silently widened to float64. Matched the existing
type: ignore[assignment] convention for numba-dispatcher/numpy-fallback
aliases in spatial.py/vicsek_predator.py, and type: ignore[no-redef] for
_kernels.py's no-numba njit shim. Dropped two now-stale type: ignore
comments in mpl_recorder.py (matplotlib stubs type view_init correctly).

engine.py's numba.config.FASTMATH = True set an attribute that doesn't
exist in numba and was never read anywhere, so perf.fastmath never
actually enabled numba fastmath compilation. Replaced it with real
per-kernel dispatch: _kernels.py now compiles fastmath=True/False
variants of each numba kernel, and spatial.py/vicsek_predator.py's
existing use_numba dispatch picks the fastmath build when
config.perf.fastmath is set.

+++ pymurmur/physics/boid_init.py
@@ -84,7 +84,7 @@ def init_positions(
         # cbrt of uniform [0,1] gives the correct radial CDF.
         R = 0.4 * min(width, height, depth)
         r = rng.uniform(0, 1, (n, 1)).astype(np.float32)
-        r = np.cbrt(r) * R  # ∛-law: uniform in volume
+        r = (np.cbrt(r) * R).astype(np.float32)  # ∛-law: uniform in volume
         dirs = random_unit_sphere(n, rng)
         return (C + dirs * r).astype(np.float32)
